% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.to.R
\name{folder.to.R}
\alias{folder.to.R}
\title{folder.to.R}
\usage{
folder.to.R(project_directory)
}
\arguments{
\item{project_directory}{...}
}
\description{
All R scripts in the folder Rfolder will be copied over to the folder R. The copied files will replaced existing files in the folder R if they share the same name.
}
\details{
To help with package organization
}
